拆掉 phase_contract 里那套已判定是错的一维遗留投影
core/phase_contract.py 涨到 1550 行（警戒线 1000），而且它同时装着两套东西：

    PhasePosture    8 字段，三锚点串在一条 depth 轴上   ← 已判定是错的那一版
    RenderPosture  28 字段，双轴 + 感知 + 执行链 + 推演  ← 正在建的这一版

问题不是行数，是**读这个文件的人很容易照着那 8 个字段写新代码**——而它们正是
被判定错了的那 8 个（遗留那套自己的注释就写着"旧，仅供既有覆盖层"）。

拆成三个，依赖单向
------------------
    core/continuum_readout.py       只读取数口（last_continuum_posture + clamp）
        ↑
    core/phase_posture_legacy.py    一维遗留投影
        ↑
    core/phase_contract.py          双轴忠实契约 + 再导出遗留名

    1550  →  1336 + 248 + 63

为什么要单独拆出取数口
----------------------
last_continuum_posture 两套契约都用。留在其中任何一边，另一边就得反向依赖，
拆开时必然循环导入；各写一份则会漂。它本身也确实是一件独立的事：
**在不构造任何东西的前提下，读进程里已经存在的那份状态**。

为什么还要再导出
----------------
拆分的目的是让读这个文件的人只看见该消费的那一份，不是去制造一次调用方大迁移。
既有 `from core.phase_contract import PhasePosture / resolve_phase_posture` 照常可用。
再导出的是**同一个对象**（有测试钉住 `is` 关系），不是各写一份。

判据：行为零变化
----------------
重跑 scripts/gen_ts_types.py，三个 .gen.ts **逐字节相同**（md5 -c 全 OK）。
panel 重建后 dist/ 也无变化。这条判据比"测试都过"更强——生成器读的就是这两份
schema 描述，输出不变即说明契约本身一个字都没动。

防回退
------
tests/test_phase_contract_continuity.py 新增 6 条：
· 遗留形状确实在自己的模块里
· 再导出是同一对象（不是副本）
· phase_contract 里**不再有那些定义**（只加个新模块、原处照旧定义 = 没拆）
· 共享取数口只有一份（三处 `is` 同一个函数）
· 依赖单向、无环——只看真正的 import 语句，不扫全文：两个被拆出的模块的文档
  都正当地提到了 core.phase_contract（在解释为什么这么拆），扫全文会把说明当依赖
· 「绝不构造」这条纪律没在搬家中丢掉（去掉文档只看代码，断言没有 get_openclawd 调用）

基线
----
file_complexity 基线里 phase_contract 从 1550 收紧到 1336（不是抬高，是收回）。

验证
----
· 相关 2634 条全过；test_phase_contract_continuity 21 条全过
· flake8 0 / black / isort 全绿
· 四道结构守卫全绿
· 三个 .gen.ts 逐字节相同；panel tsc + build 通过，dist 无变化
