+++ config/file_complexity_baseline.json
@@ -82,7 +82,7 @@
     "core/operational_slo_metrics.py": 1392,
     "core/operator_surface.py": 2169,
     "core/orchestration_authority/legacy_paths.py": 1891,
-    "core/phase_contract.py": 1550,
+    "core/phase_contract.py": 1336,
     "core/platform_closure_audit.py": 1286,
     "core/policy/alignment_surface.py": 1290,
     "core/pr3_session_continuity_authority.py": 1103,
